Renamed navigation to main-navigation
