+++ src/app/(home)/layout.tsx
@@ -1,5 +1,5 @@
 import { type PropsWithChildren } from "react";
-import { Navigation } from "~/app/components/navigation/navigation";
+import { Navigation } from "~/app/components/navigation/main-navigation";
 import { Footer } from "~/app/components/navigation/footer";
 
 export default function Layout({ children }: PropsWithChildren) {
